extract account_tx_version_detector.py
